arrumando bugs

+++ app/Http/Controllers/AlunoController.php
@@ -1,10 +1,11 @@
 <?php
-
+ 
 namespace App\Http\Controllers;
-
+ 
 use App\Models\Aluno;
+use App\Http\Requests\AlunoRequest;
 use Illuminate\Http\Request;
-
+ 
 class AlunoController extends Controller
 {
     public function index()
@@ -12,41 +13,48 @@ class AlunoController extends Controller
         $alunos = Aluno::all();
         return view('alunos.index', compact('alunos'));
     }
-
+ 
     public function show($id)
     {
         $aluno = Aluno::findOrFail($id);
         return view('alunos.show', compact('aluno'));
     }
-
+ 
     public function create()
     {
+        $this->authorize('create', Aluno::class);
         return view('alunos.create');
     }
-
+ 
     public function store(AlunoRequest $request)
-{
-    Aluno::create($request->validated());
-    return redirect()->route('alunos.index');
-}
-
+    {
+        $this->authorize('create', Aluno::class);
+        Aluno::create($request->validated());
+        return redirect()->route('alunos.index');
+    }
+ 
     public function edit($id)
     {
         $aluno = Aluno::findOrFail($id);
+        $this->authorize('update', $aluno);
         return view('alunos.edit', compact('aluno'));
     }
-
-    public function update(Request $request, $id)
+ 
+    public function update(AlunoRequest $request, $id)
     {
         $aluno = Aluno::findOrFail($id);
-        $aluno->update($request->all());
+        $this->authorize('update', $aluno);
+        $aluno->update($request->validated());
         return redirect()->route('alunos.show', $aluno->id);
     }
-
+ 
     public function destroy($id)
     {
+        $aluno = Aluno::findOrFail($id);
+        $this->authorize('delete', $aluno);
         Aluno::destroy($id);
         return redirect()->route('alunos.index');
     }
 }
-
+ 
+
